database/requirements.txt
paho.mqtt == 1.6.1
Flask == 2.2.2
boto3 >= 1.26.100
